+++ R/pas_getLabels.R
@@ -4,8 +4,6 @@
 #' @title Return labels from filtered PurpleAir Synoptic objects
 #' 
 #' @param pas PurpleAir Synoptic \emph{pas} object.
-#' @param countryCodes ISO country codes used to subset the data.
-#' @param stateCodes ISO country codes used to subset the data.
 #' @param pattern Text pattern used to filter station labels.
 #' @param isOutside Logical, is the sensor located outside?
 #' @param isParent Logigal, is the sensor a parent station?
@@ -23,8 +21,6 @@
 #' 
 pas_getLabels <- function(
   pas = NULL,
-  countryCodes = c('US'),
-  stateCodes = NULL,
   pattern = ".*",
   isOutside = TRUE,
   isParent = TRUE
@@ -39,8 +35,6 @@ pas_getLabels <- function(
   labels <- pas_getColumn(
     pas, 
     name = "label", 
-    countryCodes = countryCodes,
-    stateCodes = stateCodes, 
     pattern = pattern, 
     isOutside = isOutside, 
     isParent = isParent