+++ NEWS.md
@@ -1,3 +1,11 @@
+# AirSensor 0.6.8
+
+* Changed default value to `countryCodes = NULL` in `pas_createNew()` and
+`enhanceSynopticData()` so that it is easier to create _pas_ objects for other
+countries.
+* Removed `countryCode` and `stateCode` arguments from `pas_get~` functions.
+Filter should be done with `pas_filter()` before these functions are called.
+
 # AirSensor 0.6.7
 
 * Fixed bug in `enhanceSynopticData()` which failed to add `pwfsl~` columns when